+++ tests/test_simulation_validation.py
@@ -6,7 +6,6 @@ from pathlib import Path
 from unittest.mock import MagicMock, Mock, call, patch
 
 import numpy.testing as npt
-import pandas as pd
 import pytest
 
 import bluepysnap.simulation_validation as test_module
@@ -294,16 +293,11 @@ def test_validate_connection_overrides(mock_validate_override):
 
 
 def test__get_ids_from_spike_file(tmp_path):
-    spike_path = tmp_path / "spikes.dat"
-    pd.DataFrame({"/scatter": [1]}).to_csv(spike_path, sep="\t")
-
-    assert test_module._get_ids_from_spike_file(spike_path) == {0}
-
     spike_path = TEST_DATA_DIR / "input_spikes.h5"
     assert test_module._get_ids_from_spike_file(spike_path) == {"default": {0}}
 
-    with pytest.raises(IOError, match=r"Unknown file type: '.fake' \(supported: '.h5', '.dat'\)"):
-        test_module._get_ids_from_spike_file("fake_spikes.fake")
+    with pytest.raises(IOError, match=r"Unsupported file type: '.dat' \(supported: '.h5'\)"):
+        test_module._get_ids_from_spike_file(tmp_path / "spikes.dat")
 
 
 def test__get_ids_from_node_set():
@@ -321,6 +315,22 @@ def test__get_ids_from_node_set():
         {"fake_node_set": {"fake_prop": "fake_value"}}
     )
     assert test_module._get_ids_from_node_set("fake_node_set", config) == {}
+
+
+def test__get_ids_from_populations():
+    with copy_test_data() as (_, config_path):
+        with edit_config(config_path) as circuit_config:
+            circuit_config["networks"]["nodes"][0]["populations"]["default2"]["type"] = "virtual"
+
+        config = {"_circuit_config": TEST_DATA_DIR / "circuit_config.json"}
+        res = test_module._get_ids_from_populations(config)
+        expected = {"default": [0, 1, 2], "default2": [0, 1, 2, 3]}
+        npt.assert_equal(res, expected)
+
+        config = {"_circuit_config": config_path}
+        res = test_module._get_ids_from_populations(config, only_non_virtual=True)
+        expected = {"default": [0, 1, 2]}
+        npt.assert_equal(res, expected)
 
 
 def test__get_missing_ids():
@@ -367,37 +377,59 @@ def test__compare_ids(mock_missing_ids):
     assert test_module._compare_ids(None, None, source, prefix) == expected
 
 
-@patch.object(test_module, "_get_ids_from_node_set", new=Mock())
-@patch.object(test_module, "_resolve_path", new=Mock())
 @patch.object(test_module, "_get_ids_from_spike_file")
-@patch.object(test_module, "_get_missing_ids")
-def test__validate_spike_file_contents(mock_missing_ids, mock_ids_from_spikes):
-    input_config = {"source": "fake_node_set", "spike_file": "fake_spikes.h5"}
+@pytest.mark.parametrize(
+    "input_config,expected_message,spike_ids_side_effect",
+    [
+        [{"spike_file": "fake_spikes.h5"}, None, lambda *_: {0, 1}],
+        [
+            {"source": "fake_node_set", "spike_file": "fake_spikes.h5"},
+            "3 id(s) not found in node set 'fake_node_set': 5, 6, 7",
+            lambda *_: {5, 6, 7},
+        ],
+        [
+            {"spike_file": "fake_spikes.h5"},
+            (
+                "3 id(s) not found in node populations: "
+                "('fake_population', 0), ('fake_population', 1), ('fake_population', 2)"
+            ),
+            lambda *_: {"fake_population": [0, 1, 2]},
+        ],
+        [{"source": "fake_node_set", "spike_file": "fake_spikes.h5"}, None, lambda *_: {0, 1}],
+        [
+            {"spike_file": "fake_spikes.h5"},
+            "3 id(s) not found in node populations: 5, 6, 7",
+            lambda *_: {5, 6, 7},
+        ],
+        [
+            {"source": "fake_node_set", "spike_file": "fake_spikes.h5"},
+            (
+                "3 id(s) not found in node set 'fake_node_set': "
+                "('fake_population', 0), ('fake_population', 1), ('fake_population', 2)"
+            ),
+            lambda *_: {"fake_population": [0, 1, 2]},
+        ],
+        [{"spike_file": "fake_spikes.h5"}, "Unknown IOError", IOError("Unknown", "IOError")],
+    ],
+)
+def test__validate_spike_file_contents(
+    mock_ids_from_spikes, input_config, expected_message, spike_ids_side_effect
+):
+    config = {
+        "_config_dir": TEST_DATA_DIR,
+        "_circuit_config": TEST_DATA_DIR / "circuit_config.json",
+        "_node_sets_instance": NodeSets.from_dict(
+            {"fake_node_set": {"population": ["default"], "node_id": [0, 1]}}
+        ),
+    }
+    prefix = "fake"
+    mock_ids_from_spikes.side_effect = spike_ids_side_effect
+    if expected_message is not None:
+        expected = [BluepySnapValidationError.fatal(f"{prefix}: {expected_message}")]
+    else:
+        expected = []
+    res = test_module._validate_spike_file_contents(input_config, config, prefix)
 
-    mock_missing_ids.return_value = []
-    res = test_module._validate_spike_file_contents(input_config, config=None, prefix="")
-    expected = []
-    assert res == expected
-
-    mock_missing_ids.return_value = [0, 1, 2]
-    res = test_module._validate_spike_file_contents(input_config, config=None, prefix="fake_prefix")
-    msg = "fake_prefix: 3 id(s) not found in node set 'fake_node_set': 0, 1, 2"
-    expected = [BluepySnapValidationError.fatal(msg)]
-    assert res == expected
-
-    mock_missing_ids.return_value = [("fake_population", id_) for id_ in [0, 1, 2]]
-    res = test_module._validate_spike_file_contents(input_config, config=None, prefix="fake_prefix")
-    msg = (
-        "fake_prefix: 3 id(s) not found in node set 'fake_node_set': "
-        "('fake_population', 0), ('fake_population', 1), ('fake_population', 2)"
-    )
-    expected = [BluepySnapValidationError.fatal(msg)]
-    assert res == expected
-
-    mock_ids_from_spikes.side_effect = IOError("Unknown", "IOError")
-    res = test_module._validate_spike_file_contents(input_config, config=None, prefix="fake_prefix")
-    msg = "fake_prefix: Unknown IOError"
-    expected = [BluepySnapValidationError.fatal(msg)]
     assert res == expected
 
 
@@ -405,7 +437,6 @@ def test__validate_spike_input():
     node_sets = NodeSets.from_dict({"fake_node_set": {"node_id": [0]}})
 
     input_config = {
-        "source": "fake_node_set",
         "spike_file": TEST_DATA_DIR / "input_spikes.h5",
     }
     config = {
@@ -421,7 +452,6 @@ def test__validate_spike_input():
     }
 
     expected_error_messages = [
-        "inputs.test.source: Unknown node set: 'fail_node_set'",
         f"inputs.test.spike_file: No such file: {input_config['spike_file']}",
         "inputs.test.spike_file: Can not validate file contents",
     ]
@@ -535,16 +565,14 @@ def test_validate_inputs():
             "pass_3": {"module": "not_synapse_replay", "source": "fail_node_set"},
             "pass_4": {"module": "not_synapse_replay", "spike_file": fail_spike_file},
             "fail_0": {"module": "test_module", "node_set": "fail_node_set"},
-            "fail_1": {"module": "synapse_replay", "source": "fail_node_set"},
-            "fail_2": {"module": "synapse_replay", "spike_file": fail_spike_file},
+            "fail_1": {"module": "synapse_replay", "spike_file": fail_spike_file},
         },
     }
 
     expected_error_messages = [
         "inputs.fail_0.node_set: Unknown node set: 'fail_node_set'",
-        "inputs.fail_1.source: Unknown node set: 'fail_node_set'",
-        f"inputs.fail_2.spike_file: No such file: {fail_spike_file}",
-        f"inputs.fail_2.spike_file: Can not validate file contents",
+        f"inputs.fail_1.spike_file: No such file: {fail_spike_file}",
+        f"inputs.fail_1.spike_file: Can not validate file contents",
     ]
 
     expected = [BluepySnapValidationError.fatal(msg) for msg in expected_error_messages]
@@ -760,27 +788,12 @@ def test_validate_reports(tmp_path):
     assert test_module.validate_reports(config) == expected
 
 
-def test__get_ids_from_non_virtual_pops():
-    config = {"_circuit_config": TEST_DATA_DIR / "circuit_config.json"}
-    res = test_module._get_ids_from_non_virtual_pops(config)
-    expected = {"default": [0, 1, 2], "default2": [0, 1, 2, 3]}
-    npt.assert_equal(res, expected)
-
-    with copy_test_data() as (_, config_path):
-        with edit_config(config_path) as circuit_config:
-            circuit_config["networks"]["nodes"][0]["populations"]["default2"]["type"] = "virtual"
-
-        config = {"_circuit_config": config_path}
-        res = test_module._get_ids_from_non_virtual_pops(config)
-        expected = {"default": [0, 1, 2]}
-        npt.assert_equal(res, expected)
-
-
 def test__validate_electrodes_file():
     path = "./fake_path"
+    prefix = "run.electrodes_file"
     expected = [
-        BluepySnapValidationError.fatal(f"run.electrodes_file: No such file: {TEST_DATA_DIR/path}"),
-        BluepySnapValidationError.fatal(f"run.electrodes_file: Can not validate file contents"),
+        BluepySnapValidationError.fatal(f"{prefix}: No such file: {TEST_DATA_DIR/path}"),
+        BluepySnapValidationError.fatal(f"{prefix}: Can not validate file contents"),
     ]
     config = {"run": {"electrodes_file": path}, "_config_dir": TEST_DATA_DIR}
     assert test_module.validate_run(config) == expected
@@ -793,9 +806,9 @@ def test__validate_electrodes_file():
     }
     assert test_module.validate_run(config) == []
 
-    with patch.object(test_module, "_get_ids_from_non_virtual_pops") as patched:
+    with patch.object(test_module, "_get_ids_from_populations") as patched:
         patched.return_value = {"default": {0}}
-        msg = "run.electrodes_file: 1 id(s) not found in non-virtual populations: ('default', 1)"
+        msg = f"{prefix}: 1 id(s) not found in non-virtual node populations: ('default', 1)"
         expected = [BluepySnapValidationError.fatal(msg)]
         assert test_module.validate_run(config) == expected
 
@@ -804,12 +817,12 @@ def test__validate_electrodes_file():
     assert test_module.validate_run(config) == []
 
     config["node_set"] = "Layer23"
-    msg = "run.electrodes_file: 1 id(s) not found in node set 'Layer23': ('default', 1)"
+    msg = f"{prefix}: 1 id(s) not found in node set 'Layer23': ('default', 1)"
     expected = [BluepySnapValidationError.fatal(msg)]
     assert test_module.validate_run(config) == expected
 
     config["_circuit_config"] = ""
-    msg = "run.electrodes_file: Can not validate file contents"
+    msg = f"{prefix}: Can not validate file contents"
     expected = [BluepySnapValidationError.fatal(msg)]
     assert test_module.validate_run(config) == expected